adyam/requirements.txt
django
djangorestframework
markdown
pygments
django-filter
django-guardian